Large refactor - Improved durability of client
The client was continuously being flaky so I set out to add some more
meaningful tests and break things as much as possible. We finally have a
client and proxy that has a reasonably high chance of success. There are
some really rough edges in the proxy, which is only implemented in the
XioServerFunctionTest, but I will soften those in up coming
commits.