Add HttpStreamDecoder
The HttpStreamDecoder is a best effort HttpObjectAggregator that will
revert to streaming HttpObjects if the message to message decoding
fails.  This is mainly meant for cases where smaller messages can be
allow for optional features (md5 of entire message) but larger messages
should still be streamed, something that HttpObjectAggregator doesn't allow.